Fixed delete action modalw

+++ src/app/settings/settings.component.ts
@@ -80,16 +80,13 @@ export class SettingsComponent implements OnInit {
   }
 
   deleteAction(id) {
-    console.log(id)
-    this.actiontypedataService.delete(id)
-      .subscribe(
-      request => { location.reload() },
-      error => {
-        window.alert(error)
-      }
-      );
-    location.reload()
-
+    if (window.confirm("Weet je zeker dat je deze actie wilt verwijderen?")) {
+      this.actiontypedataService.delete(id)
+        .subscribe(
+        request => { this.actiontypedataService.getAll().subscribe(request => this.actiontypes = request) },
+        error => { window.alert("Actie kan niet verwijderd worden omdat deze in gebruik is") }
+        );
+    }
   }
 
   deleteStatus(id) {
@@ -97,7 +94,7 @@ export class SettingsComponent implements OnInit {
       this.statusDataService.delete(id)
         .subscribe(
         request => { this.statusDataService.getAll().subscribe(request => this.statuses = request); },
-        error => { window.alert("Actie kan niet verwijderd worden omdat deze in gebruik is") }
+        error => { window.alert("Status kan niet verwijderd worden omdat deze in gebruik is") }
         );
     }
   }
